move Map files into separate directory
