+++ src/pages/map/CategoryPicker.js
