+++ src/pages/map/Detail.test.js
