+++ src/pages/map/SubmitTipFooter.js
